# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Online Bookstore API
  description: API for managing a simple online bookstore, including books, shopping cart, and order history.
  version: 1.0.0
servers:
  - url: http://localhost:8811
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        coverImage:
          type: string
    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        book:
          $ref: '#/components/schemas/Book'
        quantity:
          type: integer
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        book:
          $ref: '#/components/schemas/Book'
        quantity:
          type: integer
        priceAtTimeOfPurchase:
          type: number
          format: double
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderDate:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    CartItemRequest:
      type: object
      required:
        - bookId
        - quantity
      properties:
        bookId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    UpdateQuantityRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
          minimum: 1

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Username already taken or invalid data

  /api/auth/login:
    post:
      summary: Login to get a JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  type:
                    type: string
                    example: Bearer
        '401':
          description: Invalid credentials

  /api/books:
    get:
      summary: Get all books
      tags:
        - Books
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

  /api/books/{id}:
    get:
      summary: Get book details by ID
      tags:
        - Books
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found

  /api/cart:
    get:
      summary: Get current user's cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of items in cart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
    post:
      summary: Add item to cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequest'
      responses:
        '201':
          description: Item added successfully
        '404':
          description: Book not found
    delete:
      summary: Clear entire cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart cleared successfully

  /api/cart/{itemId}:
    put:
      summary: Update cart item quantity
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuantityRequest'
      responses:
        '200':
          description: Quantity updated successfully
        '404':
          description: Cart item not found
    delete:
      summary: Remove item from cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Item removed successfully
        '404':
          description: Cart item not found

  /api/cart/checkout:
    post:
      summary: Process checkout and create order
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order placed successfully
        '400':
          description: Empty cart

  /api/orders:
    get:
      summary: Get user's order history
      tags:
        - Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'